services/api/pyproject.toml
[project]
name = "voyagent-api"
version = "0.0.0"
description = "Voyagent public HTTP + SSE API (FastAPI)."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "sse-starlette>=2.1",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.9",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "redis[hiredis]>=5.0",
    "argon2-cffi>=23.1",
    "email-validator>=2.1",
    "cryptography>=42",
    "voyagent-schemas-canonical",
    "voyagent-schemas-storage",
    "voyagent-agent-runtime",
]

[tool.uv.sources.voyagent-schemas-storage]
workspace = true

[project.scripts]
voyagent-api = "voyagent_api.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voyagent_api"]

[tool.uv.sources]
voyagent-schemas-canonical = { workspace = true }
voyagent-agent-runtime = { workspace = true }
